requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
django-cors-headers==3.5.0
django-3-jet==1.0.8
Django==3.1.2
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.1
gunicorn==20.0.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
drf-renderer-xlsx==0.3.8
djoser==2.1.0
